feat: extract script execution logic to script_engine and improve control flow parsing
